requirements.txt
torch
torchvision
streamlit
tqdm
torchaudio
pillow
